requirements/installed-locally-requirements.txt
altgraph==0.17
appdirs==1.4.4
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
distlib==0.3.1
Django==3.1
django-bootstrap3==14.2.0
django-braces==1.14.0
django-webpack-loader==0.7.0
docutils==0.16
Faker==4.14.2
ffpyplayer==4.3.1
filelock==3.0.12
future==0.18.2
idna==2.10
importlib-metadata==1.7.0
Kivy==1.11.1
kivy-deps.angle==0.1.10
kivy-deps.glew==0.1.12
kivy-deps.gstreamer==0.1.18
kivy-deps.sdl2==0.1.23
Kivy-Garden==0.1.4
misaka==2.1.1
pefile==2019.4.18
Pillow==7.2.0
pycparser==2.20
Pygments==2.6.1
pyinstaller==4.0
pyinstaller-hooks-contrib==2020.7
pypiwin32==223
python-dateutil==2.8.1
pytz==2020.1
pywin32==228
pywin32-ctypes==0.2.0
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
urllib3==1.25.10
virtualenv==20.0.30
zipp==3.1.0
